Convert a sfnetwork into a linnet

A method to convert an object of class sfnetwork into linnet format and enhance the interoperability between sfnetworks and spatstat. Use this method without the .sfnetwork suffix and after loading the spatstat package.

Usage

as.linnet.sfnetwork(X, ...)

Arguments

X

An object of class sfnetwork with a projected CRS.

...

Arguments passed to linnet.

Value

An object of class linnet.

See also