sf methods for sfnetwork objects

Tidyverse methods for sf objects. Use these methods without the .sfnetwork suffix and after loading the sf package.

# S3 method for sfnetwork
st_as_sf(x, ...)

# S3 method for sfnetwork
st_crs(x, ...)

# S3 method for sfnetwork
st_crs(x) <- value

# S3 method for sfnetwork
st_shift_longitude(x, ...)

# S3 method for sfnetwork
st_transform(x, crs, ...)

# S3 method for sfnetwork
st_wrap_dateline(x, ...)

st_dimension.sfnetwork(x, ...)

st_is_empty.sfnetwork(x)

st_is_simple.sfnetwork(x)

st_is_valid.sfnetwork(x, ...)

st_contains.sfnetwork(x, y = x, ...)

st_contains_properly.sfnetwork(x, y = x, ...)

st_covers.sfnetwork(x, y = x, ...)

st_covered_by.sfnetwork(x, y = x, ...)

st_crosses.sfnetwork(x, y = x, ...)

st_disjoint.sfnetwork(x, y = x, ...)

st_equals.sfnetwork(x, y = x, ...)

st_equals_exact.sfnetwork(x, y = x, ...)

# S3 method for sfnetwork
st_intersects(x, y = x, ...)

st_is_within_distance.sfnetwork(x, y = x, ...)

st_overlaps.sfnetwork(x, y = x, ...)

st_touches.sfnetwork(x, y = x, ...)

st_within.sfnetwork(x, y = x, ...)

# S3 method for sfnetwork
st_area(x)

st_distance.sfnetwork(x, y = x, ...)

st_length.sfnetwork(x)

# S3 method for sfnetwork
st_reverse(x, ...)

# S3 method for sfnetwork
st_simplify(x, ...)

# S3 method for sfnetwork
st_bbox(x, ...)

# S3 method for sfnetwork
st_coordinates(x, ...)

# S3 method for sfnetwork
st_geometry(x, ...)

# S3 method for sfnetwork
st_geometry(x) <- value

st_geometry_type.sfnetwork(x, ...)

# S3 method for sfnetwork
st_is(x, ...)

# S3 method for sfnetwork
st_crop(x, y, ...)

# S3 method for sfnetwork
st_filter(x, y, ..., .predicate = st_intersects)

# S3 method for sfnetwork
st_join(x, y, join = st_intersects, ..., left = TRUE)

st_nearest_feature.sfnetwork(x, y)

st_relate.sfnetwork(x, y, ...)

Arguments

x
An object of class sfnetwork.

...
Arguments passed on the corresponding sf function.

value
See st_crs or st_geometry.

crs
See st_transform.

y
An object of class sfnetwork or class sf. In some cases, it can also be an object of sfc, sfg or bbox. Always look at the documentation of the corresponding sf function for details.

.predicate
See st_filter.

join
See st_join.

left
See st_join.

Details

See the sf documentation for details on all functions.