sfnetworks: Tidy Geospatial Networks

Provides a tidy approach to spatial network analysis, in the form of classes and functions that enable a seamless interaction between the network analysis package 'tidygraph' and the spatial analysis package 'sf'.

See also

Author

Maintainer: Lucas van der Meer luukvandermeer@live.nl (ORCID)

Authors:

Lorena Abad lore.abad6@gmail.com (ORCID)

Andrea Gilardi andrea.gilardi@unimib.it (ORCID)

Robin Lovelace r.lovelace@leeds.ac.uk (ORCID)