Query spatial edge measures

These functions are a collection of specific spatial edge measures, that form a spatial extension to edge measures in tidygraph.

edge_azimuth()

edge_circuity(Inf_as_NaN = TRUE)

edge_length()

edge_displacement()

Arguments

Inf_as_NaN
Should the circuity values of loop edges (which are infinite since the straight-line distance between the boundary nodes of loop edges is 0) be stored as NaN instead of Inf? Defaults to TRUE.

Value

A numeric vector of the same length as the number of edges in the graph.

Details

Just as with all query functions in tidygraph, spatial edge measures are meant to be called inside tidygraph verbs such as mutate or filter, where the network that is currently being worked on is known and thus not needed as an argument to the function. If you want to use an algorithm outside of the tidygraph framework you can use with_graph to set the context temporarily while the algorithm is being evaluated.

Functions

edge_azimuth: The angle in radians between a straight line from the edge startpoint pointing north, and the straight line from the edge startpoint and the edge endpoint. Calculated with st_geod_azimuth. Requires a geographic CRS.

edge_circuity: The ratio of the length of an edge linestring geometry versus the straight-line distance between its boundary nodes, as described in Giacomin & Levinson, 2015.

edge_length: The length of an edge linestring geometry as calculated by st_length.

edge_displacement: The straight-line distance between the two boundary nodes of an edge, as calculated by st_distance.

Examples

library(sf, quietly = TRUE)
library(tidygraph, quietly = TRUE)

net = as_sfnetwork(roxel)

net %>%
  activate("edges") %>%
  mutate(azimuth = edge_azimuth())

#> # A sfnetwork with 701 nodes and 851 edges
#> #
#> # CRS:  EPSG:4326 
#> #
#> # A directed multigraph with 14 components with spatially explicit edges
#> #
#> # Edge Data:     851 x 6 (active)
#> # Geometry type: LINESTRING
#> # Dimension:     XY
#> # Bounding box:  xmin: 7.522594 ymin: 51.94151 xmax: 7.546705 ymax: 51.9612
#>    from    to name       type                                  geometry  azimuth
#>   <int> <int> <chr>      <fct>                         <LINESTRING [°]>    [rad]
#> 1     1     2 Havixbeck… reside… (7.533722 51.95556, 7.533461 51.95576) -0.6700…
#> 2     3     4 Pienersal… second… (7.532442 51.95422, 7.53236 51.95377,… -2.9135…
#> 3     5     6 Schulte-B… reside… (7.532709 51.95209, 7.532823 51.95239…  0.2032…
#> 4     7     8 NA         path    (7.540063 51.94468, 7.539696 51.94479… -0.9676…
#> 5     9    10 Welsinghe… reside…  (7.537673 51.9475, 7.537614 51.94562) -3.1221…
#> 6    11    12 NA         footway (7.543791 51.94733, 7.54369 51.94686,… -3.0971…
#> # … with 845 more rows
#> #
#> # Node Data:     701 x 1
#> # Geometry type: POINT
#> # Dimension:     XY
#> # Bounding box:  xmin: 7.522622 ymin: 51.94151 xmax: 7.546705 ymax: 51.9612
#>              geometry
#>           <POINT [°]>
#> 1 (7.533722 51.95556)
#> 2 (7.533461 51.95576)
#> 3 (7.532442 51.95422)
#> # … with 698 more rows

net %>%
  activate("edges") %>%
  mutate(circuity = edge_circuity())

#> # A sfnetwork with 701 nodes and 851 edges
#> #
#> # CRS:  EPSG:4326 
#> #
#> # A directed multigraph with 14 components with spatially explicit edges
#> #
#> # Edge Data:     851 x 6 (active)
#> # Geometry type: LINESTRING
#> # Dimension:     XY
#> # Bounding box:  xmin: 7.522594 ymin: 51.94151 xmax: 7.546705 ymax: 51.9612
#>    from    to name       type                                  geometry circuity
#>   <int> <int> <chr>      <fct>                         <LINESTRING [°]>      [1]
#> 1     1     2 Havixbeck… resid…  (7.533722 51.95556, 7.533461 51.95576) 1.000000
#> 2     3     4 Pienersal… secon… (7.532442 51.95422, 7.53236 51.95377, … 1.005896
#> 3     5     6 Schulte-B… resid… (7.532709 51.95209, 7.532823 51.95239,… 1.000472
#> 4     7     8 NA         path   (7.540063 51.94468, 7.539696 51.94479,… 1.008539
#> 5     9    10 Welsinghe… resid…   (7.537673 51.9475, 7.537614 51.94562) 1.000000
#> 6    11    12 NA         footw… (7.543791 51.94733, 7.54369 51.94686, … 1.021405
#> # … with 845 more rows
#> #
#> # Node Data:     701 x 1
#> # Geometry type: POINT
#> # Dimension:     XY
#> # Bounding box:  xmin: 7.522622 ymin: 51.94151 xmax: 7.546705 ymax: 51.9612
#>              geometry
#>           <POINT [°]>
#> 1 (7.533722 51.95556)
#> 2 (7.533461 51.95576)
#> 3 (7.532442 51.95422)
#> # … with 698 more rows

net %>%
  activate("edges") %>%
  mutate(length = edge_length())

#> # A sfnetwork with 701 nodes and 851 edges
#> #
#> # CRS:  EPSG:4326 
#> #
#> # A directed multigraph with 14 components with spatially explicit edges
#> #
#> # Edge Data:     851 x 6 (active)
#> # Geometry type: LINESTRING
#> # Dimension:     XY
#> # Bounding box:  xmin: 7.522594 ymin: 51.94151 xmax: 7.546705 ymax: 51.9612
#>    from    to name       type                                  geometry   length
#>   <int> <int> <chr>      <fct>                         <LINESTRING [°]>      [m]
#> 1     1     2 Havixbeck… reside… (7.533722 51.95556, 7.533461 51.95576)  28.859…
#> 2     3     4 Pienersal… second… (7.532442 51.95422, 7.53236 51.95377,… 107.704…
#> 3     5     6 Schulte-B… reside… (7.532709 51.95209, 7.532823 51.95239…  54.362…
#> 4     7     8 NA         path    (7.540063 51.94468, 7.539696 51.94479… 155.230…
#> 5     9    10 Welsinghe… reside…  (7.537673 51.9475, 7.537614 51.94562) 208.663…
#> 6    11    12 NA         footway (7.543791 51.94733, 7.54369 51.94686,…  63.023…
#> # … with 845 more rows
#> #
#> # Node Data:     701 x 1
#> # Geometry type: POINT
#> # Dimension:     XY
#> # Bounding box:  xmin: 7.522622 ymin: 51.94151 xmax: 7.546705 ymax: 51.9612
#>              geometry
#>           <POINT [°]>
#> 1 (7.533722 51.95556)
#> 2 (7.533461 51.95576)
#> 3 (7.532442 51.95422)
#> # … with 698 more rows

net %>%
  activate("edges") %>%
  mutate(displacement = edge_displacement())

#> # A sfnetwork with 701 nodes and 851 edges
#> #
#> # CRS:  EPSG:4326 
#> #
#> # A directed multigraph with 14 components with spatially explicit edges
#> #
#> # Edge Data:     851 x 6 (active)
#> # Geometry type: LINESTRING
#> # Dimension:     XY
#> # Bounding box:  xmin: 7.522594 ymin: 51.94151 xmax: 7.546705 ymax: 51.9612
#>    from    to name      type                               geometry displacement
#>   <int> <int> <chr>     <fct>                      <LINESTRING [°]>          [m]
#> 1     1     2 Havixbec… resid… (7.533722 51.95556, 7.533461 51.955…     28.85976
#> 2     3     4 Pienersa… secon… (7.532442 51.95422, 7.53236 51.9537…    107.07330
#> 3     5     6 Schulte-… resid… (7.532709 51.95209, 7.532823 51.952…     54.33675
#> 4     7     8 NA        path   (7.540063 51.94468, 7.539696 51.944…    153.91629
#> 5     9    10 Welsingh… resid… (7.537673 51.9475, 7.537614 51.9456…    208.66369
#> 6    11    12 NA        footw… (7.543791 51.94733, 7.54369 51.9468…     61.70258
#> # … with 845 more rows
#> #
#> # Node Data:     701 x 1
#> # Geometry type: POINT
#> # Dimension:     XY
#> # Bounding box:  xmin: 7.522622 ymin: 51.94151 xmax: 7.546705 ymax: 51.9612
#>              geometry
#>           <POINT [°]>
#> 1 (7.533722 51.95556)
#> 2 (7.533461 51.95576)
#> 3 (7.532442 51.95422)
#> # … with 698 more rows