Extract the active element of a sfnetwork as spatial tibble

The sfnetwork method for as_tibble is conceptually different. Whenever a geometry list column is present, it will by default return what we call a 'spatial tibble'. With that we mean an object of class c('sf', 'tbl_df') instead of an object of class 'tbl_df'. This little conceptual trick is essential for how tidyverse functions handle sfnetwork objects, i.e. always using the corresponding sf method if present. When using as_tibble on sfnetwork objects directly as a user, you can disable this behaviour by setting spatial = FALSE.

Usage

# S3 method for class 'sfnetwork'
as_tibble(x, active = NULL, focused = TRUE, spatial = TRUE, ...)

Arguments

x

An object of class sfnetwork.

active

Which network element (i.e. nodes or edges) to activate before extracting. If NULL, it will be set to the current active element of the given network. Defaults to NULL.

focused

Should only features that are in focus be extracted? Defaults to TRUE. See focus for more information on focused networks.

spatial

Should the extracted tibble be a 'spatial tibble', i.e. an object of class c('sf', 'tbl_df'), if it contains a geometry list column. Defaults to TRUE.

...

Arguments passed on to as_tibble.

Value

The active element of the network as an object of class sf if a geometry list column is present and spatial = TRUE, and object of class tibble otherwise.

Examples

library(tibble, quietly = TRUE)

net = as_sfnetwork(roxel)

# Extract the active network element as a spatial tibble.
as_tibble(net)
#> Simple feature collection with 987 features and 0 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 7.522595 ymin: 51.94151 xmax: 7.546705 ymax: 51.96119
#> Geodetic CRS:  WGS 84
#> # A tibble: 987 × 1
#>               geometry
#>            <POINT [°]>
#>  1 (7.538109 51.95286)
#>  2 (7.537867 51.95282)
#>  3 (7.537815 51.95867)
#>  4 (7.537015 51.95848)
#>  5 (7.533441 51.95578)
#>  6 (7.533415 51.95561)
#>  7 (7.525977 51.95283)
#>  8 (7.526581 51.95293)
#>  9 (7.532301 51.95559)
#> 10 (7.532142 51.95589)
#> # ℹ 977 more rows

# Extract any network element as a spatial tibble.
as_tibble(net, "edges")
#> Simple feature collection with 1215 features and 4 fields
#> Attribute-geometry relationships: constant (2), NA's (2)
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 7.522595 ymin: 51.94151 xmax: 7.546705 ymax: 51.96119
#> Geodetic CRS:  WGS 84
#> # A tibble: 1,215 × 5
#>     from    to name                              type                   geometry
#>    <int> <int> <chr>                             <chr>          <LINESTRING [°]>
#>  1     1     2 Hagemanns Kämpken                 resi… (7.538109 51.95286, 7.53…
#>  2     3     4 Stiegkamp                         resi… (7.537815 51.95867, 7.53…
#>  3     5     6 Havixbecker Straße                resi… (7.533441 51.95578, 7.53…
#>  4     7     8 Holzschuhmacherweg                resi… (7.525977 51.95283, 7.52…
#>  5     9    10 Annette-von-Droste-Hülshoff-Stra… seco… (7.532301 51.95559, 7.53…
#>  6    11    12 NA                                foot… (7.543404 51.94779, 7.54…
#>  7    13    14 Deermannskamp                     resi… (7.537675 51.95689, 7.53…
#>  8    15    16 Lindenstraße                      resi… (7.539105 51.9504, 7.540…
#>  9    17    18 NA                                NA    (7.534875 51.95795, 7.53…
#> 10    19    20 Annette-von-Droste-Hülshoff-Stra… seco… (7.532465 51.95424, 7.53…
#> # ℹ 1,205 more rows

# Extract the active network element as a regular tibble.
as_tibble(net, spatial = FALSE)
#> # A tibble: 987 × 1
#>               geometry
#>            <POINT [°]>
#>  1 (7.538109 51.95286)
#>  2 (7.537867 51.95282)
#>  3 (7.537815 51.95867)
#>  4 (7.537015 51.95848)
#>  5 (7.533441 51.95578)
#>  6 (7.533415 51.95561)
#>  7 (7.525977 51.95283)
#>  8 (7.526581 51.95293)
#>  9 (7.532301 51.95559)
#> 10 (7.532142 51.95589)
#> # ℹ 977 more rows