Query spatial edge measures

These functions are a collection of specific spatial edge measures, that form a spatial extension to edge measures in tidygraph.

edge_circuity()

edge_length()

edge_displacement()

Value

A numeric vector of the same length as the number of edges in the graph.

Details

Just as with all query functions in tidygraph, spatial edge measures are meant to be called inside tidygraph verbs such as mutate or filter, where the network that is currently being worked on is known and thus not needed as an argument to the function. If you want to use an algorithm outside of the tidygraph framework you can use with_graph to set the context temporarily while the algorithm is being evaluated.

Functions

edge_circuity: The ratio of the length of an edge linestring geometry versus the straight-line distance between its boundary nodes, as described in Giacomin & Levinson, 2015.

edge_length: The length of an edge linestring geometry as calculated by st_length.

edge_displacement: The straight-line distance between the two boundary nodes of an edge, as calculated by st_distance.

Note

Note that edge_is_within_distance is a wrapper around the st_is_within_distance predicate from sf. Hence, it is based on 'as-the-crow-flies' distance, and not on distances over the network.

Examples

library(sf)
library(tidygraph)

net = as_sfnetwork(roxel, directed = FALSE) %>%
  st_transform(3035)

net %>%
  activate("edges") %>%
  mutate(circuity = edge_circuity())

#> # A sfnetwork with 701 nodes and 851 edges
#> #
#> # CRS:  EPSG:3035 
#> #
#> # An undirected multigraph with 14 components with spatially explicit edges
#> #
#> # Edge Data:     851 x 6 (active)
#> # Geometry type: LINESTRING
#> # Dimension:     XY
#> # Bounding box:  xmin: 4150707 ymin: 3206375 xmax: 4152367 ymax: 3208565
#>    from    to name       type                                  geometry circuity
#>   <int> <int> <chr>      <fct>                         <LINESTRING [m]>      [1]
#> 1     1     2 Havixbeck… resid…      (4151491 3207923, 4151474 3207946) 1.000000
#> 2     3     4 Pienersal… secon… (4151398 3207777, 4151390 3207727, 415… 1.005894
#> 3     5     6 Schulte-B… resid… (4151408 3207539, 4151417 3207573, 415… 1.000471
#> 4     7     8 NA         path   (4151885 3206698, 4151861 3206711, 415… 1.008540
#> 5     9    10 Welsinghe… resid…      (4151732 3207017, 4151721 3206809) 1.000000
#> 6    11    12 NA         footw… (4152152 3206984, 4152143 3206932, 415… 1.021397
#> # … with 845 more rows
#> #
#> # Node Data:     701 x 1
#> # Geometry type: POINT
#> # Dimension:     XY
#> # Bounding box:  xmin: 4150707 ymin: 3206375 xmax: 4152367 ymax: 3208565
#>            geometry
#>         <POINT [m]>
#> 1 (4151491 3207923)
#> 2 (4151474 3207946)
#> 3 (4151398 3207777)
#> # … with 698 more rows

net %>%
  activate("edges") %>%
  mutate(length = edge_length())

#> # A sfnetwork with 701 nodes and 851 edges
#> #
#> # CRS:  EPSG:3035 
#> #
#> # An undirected multigraph with 14 components with spatially explicit edges
#> #
#> # Edge Data:     851 x 6 (active)
#> # Geometry type: LINESTRING
#> # Dimension:     XY
#> # Bounding box:  xmin: 4150707 ymin: 3206375 xmax: 4152367 ymax: 3208565
#>    from    to name       type                                  geometry   length
#>   <int> <int> <chr>      <fct>                         <LINESTRING [m]>      [m]
#> 1     1     2 Havixbeck… reside…     (4151491 3207923, 4151474 3207946)  28.859…
#> 2     3     4 Pienersal… second… (4151398 3207777, 4151390 3207727, 41… 107.714…
#> 3     5     6 Schulte-B… reside… (4151408 3207539, 4151417 3207573, 41…  54.367…
#> 4     7     8 NA         path    (4151885 3206698, 4151861 3206711, 41… 155.223…
#> 5     9    10 Welsinghe… reside…     (4151732 3207017, 4151721 3206809) 208.682…
#> 6    11    12 NA         footway (4152152 3206984, 4152143 3206932, 41…  63.028…
#> # … with 845 more rows
#> #
#> # Node Data:     701 x 1
#> # Geometry type: POINT
#> # Dimension:     XY
#> # Bounding box:  xmin: 4150707 ymin: 3206375 xmax: 4152367 ymax: 3208565
#>            geometry
#>         <POINT [m]>
#> 1 (4151491 3207923)
#> 2 (4151474 3207946)
#> 3 (4151398 3207777)
#> # … with 698 more rows

net %>%
  activate("edges") %>%
  mutate(displacement = edge_displacement())

#> # A sfnetwork with 701 nodes and 851 edges
#> #
#> # CRS:  EPSG:3035 
#> #
#> # An undirected multigraph with 14 components with spatially explicit edges
#> #
#> # Edge Data:     851 x 6 (active)
#> # Geometry type: LINESTRING
#> # Dimension:     XY
#> # Bounding box:  xmin: 4150707 ymin: 3206375 xmax: 4152367 ymax: 3208565
#>    from    to name      type                               geometry displacement
#>   <int> <int> <chr>     <fct>                      <LINESTRING [m]>          [m]
#> 1     1     2 Havixbec… resid…   (4151491 3207923, 4151474 3207946)     28.85977
#> 2     3     4 Pienersa… secon… (4151398 3207777, 4151390 3207727, …    107.08291
#> 3     5     6 Schulte-… resid… (4151408 3207539, 4151417 3207573, …     54.34168
#> 4     7     8 NA        path   (4151885 3206698, 4151861 3206711, …    153.90859
#> 5     9    10 Welsingh… resid…   (4151732 3207017, 4151721 3206809)    208.68255
#> 6    11    12 NA        footw… (4152152 3206984, 4152143 3206932, …     61.70817
#> # … with 845 more rows
#> #
#> # Node Data:     701 x 1
#> # Geometry type: POINT
#> # Dimension:     XY
#> # Bounding box:  xmin: 4150707 ymin: 3206375 xmax: 4152367 ymax: 3208565
#>            geometry
#>         <POINT [m]>
#> 1 (4151491 3207923)
#> 2 (4151474 3207946)
#> 3 (4151398 3207777)
#> # … with 698 more rows