Spatial morphers for sfnetworks

Spatial morphers form spatial add-ons to the set of morphers provided by tidygraph. These functions are not meant to be called directly. They should either be passed into morph to create a temporary alternate representation of the input network. Such an alternate representation is a list of one or more network objects. Single elements of that list can be extracted directly as a new network by passing the morpher to convert instead. Alternatively, if the morphed state contains multiple elements, all of them can be extracted together inside a tibble by passing the morpher to crystallise.

to_spatial_coordinates(x)

to_spatial_subdivision(x)

to_spatial_directed(x)

to_spatial_explicit_edges(x, ...)

to_spatial_implicit_edges(x)

to_spatial_shortest_paths(x, ...)

to_spatial_simple(x, remove_parallels = TRUE, remove_loops = TRUE)

to_spatial_smooth(x)

to_spatial_subset(x, ..., subset_by = NULL)

Arguments

x
An object of class sfnetwork.

...
Arguments to be passed on to other functions. See the description of each morpher for details.

remove_parallels
Should parallel edges be removed. Defaults to TRUE.

remove_loops
Should loops be remove. Defaults to TRUE.

subset_by
Whether to create subgraphs based on nodes or edges.

Value

Either a morphed_sfnetwork, which is a list of one or more sfnetwork objects, or a morphed_tbl_graph, which is a list of one or more tbl_graph object. See the description of each morpher for details.

Details

It also possible to create your own morphers. See the documentation of morph for the requirements for custom morphers.

Functions

to_spatial_coordinates: Store the spatial coordinates of the nodes in separate coordinate columns, instead of a sfc geometry list column. If edges are spatially explicit, the edge geometries are dropped. Returns a morphed_tbl_graph containing a single element of class tbl_graph.

to_spatial_subdivision: Construct a subdivision of the network by subdividing edges at each interior point that is equal to any other interior or boundary point in the edges table. Interior points in this sense are those points that are included in their linestring geometry feature but are not endpoints of it, while boundary points are the endpoints of the linestrings. The network is reconstructed after subdivision such that edges are connected at the points of subdivision. Returns a morphed_sfnetwork containing a single element of class sfnetwork. This morpher requires edges to be spatially explicit.

to_spatial_directed: Make a network directed in the direction given by the linestring geometries of the edges. Differs from to_directed, which makes a network directed based on the node indices given in the from and to columns. In undirected networks these indices may not correspond with the endpoints of the linestring geometries. Returns a morphed_sfnetwork containing a single element of class sfnetwork. This morpher requires edges to be spatially explicit. If not, use to_directed.

to_spatial_explicit_edges: Create linestring geometries between from and to nodes of spatially implicit edges. If the edges data can be directly converted to an object of class sf using st_as_sf, extra arguments can be provided as ... which will be forwarded to st_as_sf internally. Otherwise, straight lines will be drawn between source and target nodes of edges. Returns a morphed_sfnetwork containing a single element of class sfnetwork.

to_spatial_implicit_edges: Remove linestring geometries of spatially explicit edges. Returns a morphed_sfnetwork containing a single element of class sfnetwork.

to_spatial_shortest_paths: Limit a network to those nodes and edges that are part of the shortest path between two nodes. ... is evaluated in the same manner as st_network_paths. Returns a morphed_sfnetwork that may contain multiple elements of class sfnetwork, depending on the number of requested paths. When unmorphing only the first instance of both the node and edge data will be used, as the the same node and/or edge can be present in multiple paths.

to_spatial_simple: Remove loops and parallel edges. Returns a morphed_sfnetwork containing a single element of class sfnetwork. This morpher requires edges to be spatially explicit. If not, use to_simple.

to_spatial_smooth: Construct a smoothed version of the network by iteratively removing pseudo nodes, while preserving the connectivity of the network. In the case of directed networks, pseudo nodes are those nodes that have only one incoming and one outgoing edge. In undirected networks, pseudo nodes are those nodes that have two incident edges. Connectivity of the network is preserved by concatenating the incident edges of each removed pseudo node. Returns a morphed_sfnetwork containing a single element of class sfnetwork.

to_spatial_subset: Subset the network by applying a spatial filter, i.e. a filter on the geometry column based on a spatial predicate. ... is evaluated in the same manner as st_filter. Returns a morphed_sfnetwork containing a single element of class sfnetwork. For filters on an attribute column, use to_subgraph.

Examples

library(sf)
library(tidygraph)

net = as_sfnetwork(roxel, directed = FALSE) %>%
  st_transform(3035)

# Temporary changes with morph and unmorph.
net %>%
 activate("edges") %>%
 mutate(weight = edge_length()) %>%
 morph(to_spatial_shortest_paths, from = 1, to = 10) %>%
 mutate(in_paths = TRUE) %>%
 unmorph()

#> # A sfnetwork with 701 nodes and 851 edges
#> #
#> # CRS:  EPSG:3035 
#> #
#> # An undirected multigraph with 14 components with spatially explicit edges
#> #
#> # Edge Data:     851 x 7 (active)
#> # Geometry type: LINESTRING
#> # Dimension:     XY
#> # Bounding box:  xmin: 4150707 ymin: 3206375 xmax: 4152367 ymax: 3208565
#>    from    to name     type                           geometry   weight in_paths
#>   <int> <int> <fct>    <fct>                  <LINESTRING [m]>      [m] <lgl>   
#> 1     1     2 Havixbe… resid… (4151491 3207923, 4151474 32079…  28.859… NA      
#> 2     3     4 Pieners… secon… (4151398 3207777, 4151390 32077… 107.714… NA      
#> 3     5     6 Schulte… resid… (4151408 3207539, 4151417 32075…  54.367… NA      
#> 4     7     8 NA       path   (4151885 3206698, 4151861 32067… 155.223… NA      
#> 5     9    10 Welsing… resid… (4151732 3207017, 4151721 32068… 208.682… TRUE    
#> 6    11    12 NA       footw… (4152152 3206984, 4152143 32069…  63.028… NA      
#> # … with 845 more rows
#> #
#> # Node Data:     701 x 1
#> # Geometry type: POINT
#> # Dimension:     XY
#> # Bounding box:  xmin: 4150707 ymin: 3206375 xmax: 4152367 ymax: 3208565
#>            geometry
#>         <POINT [m]>
#> 1 (4151491 3207923)
#> 2 (4151474 3207946)
#> 3 (4151398 3207777)
#> # … with 698 more rows

# Lasting changes with convert.
net %>%
 activate("edges") %>%
 mutate(weight = edge_length()) %>%
 convert(to_spatial_shortest_paths, from = 1, to = 10)

#> # A sfnetwork with 18 nodes and 17 edges
#> #
#> # CRS:  EPSG:3035 
#> #
#> # An unrooted tree with spatially explicit edges
#> #
#> # Edge Data:     17 x 7 (active)
#> # Geometry type: LINESTRING
#> # Dimension:     XY
#> # Bounding box:  xmin: 4151491 ymin: 3206809 xmax: 4151802 ymax: 3207923
#>    from    to name   type                     geometry   weight .tidygraph_edge…
#>   <int> <int> <fct>  <fct>            <LINESTRING [m]>      [m]            <int>
#> 1     2     3 Welsi… resi… (4151732 3207017, 4151721 … 208.682…                5
#> 2     2     6 Nottu… resi… (4151732 3207017, 4151755 …  84.406…               94
#> 3     5     8 Roxel… seco… (4151531 3207830, 4151535 …  98.312…              159
#> 4     8    12 Roxel… seco… (4151519 3207865, 4151524 …  37.001…              295
#> 5     5     7 Dorff… resi… (4151620 3207798, 4151621 …  49.307…              486
#> 6     9    15 Dorff… resi… (4151786 3207133, 4151794 …  32.664…              591
#> # … with 11 more rows
#> #
#> # Node Data:     18 x 2
#> # Geometry type: POINT
#> # Dimension:     XY
#> # Bounding box:  xmin: 4151491 ymin: 3206809 xmax: 4151802 ymax: 3207923
#>            geometry .tidygraph_node_index
#>         <POINT [m]>                 <int>
#> 1 (4151491 3207923)                     1
#> 2 (4151732 3207017)                     9
#> 3 (4151721 3206809)                    10
#> # … with 15 more rows