Compute a cost matrix of a spatial network

Wrapper around distances to calculate costs of pairwise shortest paths between points in a spatial network. It allows to provide any set of geospatial point as from and to arguments. If such a geospatial point is not equal to a node in the network, it will be snapped to its nearest node before calculating costs.

st_network_cost(x, from = igraph::V(x), to = igraph::V(x), weights = NULL, ...)

Arguments

x
An object of class sfnetwork.

from
The (set of) geospatial point(s) from which the shortest paths will be calculated. Can be an object of class sf or sfc. Empty geometries will be ignored. Alternatively, it can be a numeric vector, containing the indices of the nodes from which the shortest paths will be calculated. By default, all nodes in the network are included.

to
The (set of) geospatial point(s) to which the shortest paths will be calculated. Can be an object of class sf or sfc. Empty geometries will be ignored. Alternatively, it can be a numeric vector, containing the indices of the nodes to which the shortest paths will be calculated. By default, all nodes in the network are included.

weights
The edge weights to be used in the shortest path calculation. Can be a numeric vector giving edge weights, or a column name referring to an attribute column in the edges table containing those weights. If set to NULL, the values of a column named 'weight' in the edges table will be used automatically, as long as this column is present. If set to NA, no weights are used (even if the edges have a weight column).

...
Arguments passed on to distances.

Value

An nxm numeric matrix where n is the length of the from argument, and m is the length of the to argument.

Details

See the igraph documentation.

See also

Examples

library(sf)

net = as_sfnetwork(roxel, directed = FALSE) %>%
  st_transform(3035)

p1 = st_geometry(net, "nodes")[1]
p2 = st_geometry(net, "nodes")[9]
p3 = st_sfc(p1[[1]] + st_point(c(500, 500)), crs = st_crs(p1))
p4 = st_sfc(p2[[1]] + st_point(c(-500, -500)), crs = st_crs(p2))
pts1 = c(p1, p3)
pts2 = c(p2, p4)

st_network_cost(net, pts1, pts2)

#>      [,1] [,2]
#> [1,]   16   28
#> [2,]   23   35